refactor(scripts): move executors from monorepo to proper domain
