+++ package.json
@@ -17,13 +17,13 @@
     "buildci": "lage build test lint type-check --verbose",
     "builddemo": "yarn build --to public-docsite-resources",
     "buildto": "lage build --verbose --to",
-    "buildto:lerna": "node ./scripts/monorepo/src/buildTo.js",
+    "buildto:lerna": "node ./scripts/executors/buildTo.js",
     "bundle": "lage bundle --verbose",
     "bundlesizecollect": "node ./scripts/generators/bundle-size-collect",
     "change": "beachball change --no-commit",
     "check:change": "beachball check",
     "check:modified-files": "node -r ./scripts/ts-node/register ./scripts/executors/check-for-modified-files",
-    "check:affected-package": "node ./scripts/monorepo/src/checkIfPackagesAffected.js",
+    "check:affected-package": "node ./scripts/executors/checkIfPackagesAffected.js",
     "check:installed-dependencies-versions": "satisfied --skip-invalid --ignore \"prettier|angular|lit|sass|@storybook/html|@storybook/mdx2-csf|svelte|@testing-library|vue|@cypress/react|cypress|@swc/wasm|@cactuslab/usepubsub|react-vis|@microsoft/load-themed-styles\"",
     "clean": "lage clean --verbose",
     "code-style": "lage code-style --verbose",
@@ -50,8 +50,8 @@
     "release:fluentui:patch": "node -r ./scripts/ts-node/register ./scripts/fluentui-publish publish-patch",
     "release:fluentui:post-validation": "node -r ./scripts/ts-node/register ./scripts/fluentui-publish post-publish",
     "rename-package": "node -r ./scripts/ts-node/register ./scripts/generators/rename-package.ts",
-    "run:published": "node ./scripts/monorepo/src/runPublished.js",
-    "runto:lerna": "node ./scripts/monorepo/src/runTo.js",
+    "run:published": "node ./scripts/executors/runPublished.js",
+    "runto:lerna": "node ./scripts/executors/runTo.js",
     "scrub": "node ./scripts/executors/scrub.js",
     "start": "node ./scripts/executors/start.js",
     "start:northstar": "yarn workspace @fluentui/docs start",
