+++ scripts/monorepo/src/index.d.ts
@@ -4,8 +4,9 @@ export { default as findGitRoot } from './findGitRoot';
 export { default as findRepoDeps } from './findRepoDeps';
 export { default as getAllPackageInfo } from './getAllPackageInfo';
 export { default as isConvergedPackage } from './isConvergedPackage';
-export { default as getAffectedPackages } from './getAffectedPackages';
+export { getAffectedPackages } from './getAffectedPackages';
 export { getLernaAliases } from './get-lerna-aliases';
 export { getDefaultEnvironmentVars } from './getDefaultEnvironmentVars';
 export { getProjectMetadata, workspaceRoot } from './utils';
 export * as eslintConstants from './eslint-constants';
+export { getNthCommit } from './getNthCommit';
