+++ scripts/monorepo/src/index.js
@@ -4,8 +4,9 @@ module.exports = {
   findRepoDeps: require('./findRepoDeps'),
   getAllPackageInfo: require('./getAllPackageInfo'),
   isConvergedPackage: require('./isConvergedPackage'),
-  getAffectedPackages: require('./getAffectedPackages'),
   eslintConstants: require('./eslint-constants'),
+  ...require('./getAffectedPackages'),
+  ...require('./getNthCommit'),
   ...require('./getDefaultEnvironmentVars'),
   ...require('./get-lerna-aliases'),
   ...require('./utils'),
